% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gradient.R
\name{GD_multi_dim}
\alias{GD_multi_dim}
\title{multiple linear regression}
\usage{
GD_multi_dim(x, y, learn_rate = 0.001, iter = 500, display_step = 50)
}
\arguments{
\item{x}{data matrix}

\item{y}{data matrix}

\item{learn_rate}{number}

\item{iter}{number}

\item{display_step}{number}
}
\value{
gradient descent output
}
\description{
multiple linear regression
}
\examples{
x1 <- c(137.97, 104.50, 100.00, 124.32, 79.20, 99.00, 124.00, 114.00)
x2 <- c(3, 2, 2, 3, 1, 2, 3, 2)
x <- cbind(x1, x2)
y <- c(145.00, 110.00, 93.00, 116.00, 65.32, 104.00, 118.00, 91.00)
GD_multi_dim(x, y)
GD_multi_dim(x, y, iter = 200)
GD_multi_dim(x, y, iter = 300, display_step = 25)
}
