% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gradient.R
\name{GD_one_dim}
\alias{GD_one_dim}
\title{one-dimensional linear regression}
\usage{
GD_one_dim(x, y, learn_rate = 1e-05, iter = 100, display_step = 10)
}
\arguments{
\item{x}{data vector}

\item{y}{data vector}

\item{learn_rate}{number}

\item{iter}{number}

\item{display_step}{number}
}
\value{
gradient descent output
}
\description{
one-dimensional linear regression
}
\examples{
x <- c(137.97, 104.50, 100.00, 124.32, 79.20, 99.00, 124.00, 114.00)
y <- c(145.00, 110.00, 93.00, 116.00, 65.32, 104.00, 118.00, 81.00)
GD_one_dim(x, y)
GD_one_dim(x, y, iter = 50)
GD_one_dim(x, y, iter = 30)
GD_one_dim(x, y, display_step = 5)
GD_one_dim(x, y, iter = 50, display_step = 5)
}
